Event fix

+++ src/main/java/com/fmning/service/manager/impl/EventManagerImpl.java
@@ -29,7 +29,7 @@ public class EventManagerImpl implements EventManager{
 
 	@Override
 	public int createEvent(String type, int mappingId, String title, String description, Instant startTime,
-			Instant endTime, String location, int fee, int ownerId, int ticketTemplaceId, boolean active,
+			Instant endTime, String location, double fee, int ownerId, int ticketTemplaceId, boolean active,
 			String message, int ticketBalance) {
 		Event event = new Event();
 		event.setType(type);
